[Behat] Extract Addressing transformer
