+++ etc/behat/suites/domain_order.yml
@@ -13,9 +13,9 @@ default:
                 - sylius.behat.context.setup.shipping
                 - sylius.behat.context.setup.customer
                 - sylius.behat.context.setup.order
-                - sylius.behat.context.setup.addressing
 
                 - sylius.behat.context.transform.shared_storage
+                - sylius.behat.context.transform.addressing
 
                 - sylius.behat.context.domain.order
